+++ lib/flutter_callkit_incoming.dart
@@ -111,12 +111,14 @@ class FlutterCallkitIncoming {
     return await _channel.invokeMethod("getDevicePushTokenVoIP");
   }
 
+  /// Silence CallKit events
+  static Future silenceEvents() async {
+    return await _channel.invokeMethod("silenceEvents", true);
+  }
 
-  /// Start incoming call
-  /// On iOS: start connection timer
-  /// On Android: not implemented
-  static Future startIncomingCall() async {
-    await _channel.invokeMethod("startCallIncoming");
+  /// Unsilence CallKit events
+  static Future unsilenceEvents() async {
+    return await _channel.invokeMethod("silenceEvents", false);
   }
 
   /// Request permisstion show notification for Android(13)